Bugfix: Match textages songlist logic as close as we can
This fixes a bug where certain AC-only songs would not show. This has also revealed some issues with the underlying data.

Textage does not have certain INFINITAS songs or charts within its INFINITAS song listing data, which leads them to show in this list as unique to AC. Flash Back 90s was the first example that popped out to me in writing this.

We're considering normalizing this against Konami's site or using Kamaitachi data instead.